add flip operator both forward and backward

+++ include/backend/base_array.hpp
@@ -19,8 +19,8 @@ public:
     inline void set_ptr(Dtype * ptr) {__ptr = ptr;}
 
     virtual cached_array_type compact(size_t size, 
-                                      std::vector<size_t> shape,
-                                      std::vector<size_t> strides,
+                                      std::vector<int32_t> shape,
+                                      std::vector<int32_t> strides,
                                       size_t offset)=0;
     virtual void mem_cpy(Dtype* ptr, 
                          MemCpyType mem_cpy_type)=0;
